requirements.txt
streamlit>=1.53
starlette>=0.40
uvicorn[standard]>=0.30
itsdangerous>=2.2
pypdf>=4.0
python-dotenv>=1.0
openai>=1.0
pytest>=8.0
